fix header layout
